requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
python-dotenv==1.0.1
requests==2.32.3
